feat(canary): update query filters strategy to export plips like csv
